Repositories moved in folder Reposiroties, reorganized DataBaseFacade
