+++ src/Coworking.Data/Repositories/ClanRepository.cs
@@ -1,11 +1,12 @@
-﻿using Coworking.Domain.Entities;
+﻿using Coworking.Data.Providers;
+using Coworking.Domain.Entities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Coworking.Data.Providers
+namespace Coworking.Data.Repositories
 {
     internal class ClanRepository : IRepository<Clan>
     {
