+++ src/Coworking.Data/Repositories/IRepository.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Coworking.Data.Providers
+namespace Coworking.Data.Repositories
 {
     internal interface IRepository<T>
     {
